Reverted project to 1.5.0 and made changes.
Changelogs:
- Moved tunable parameters as a card in Custom Game settings.
- Changed Icon messages to display the number and Icon.
- Changed Aura messages to display the number and color of the aura.
- Added easter egg uwu hello and thanks emotes.
- Set default ultimate rate from 500% to 100%.
- Set default ability cooldowns from 20% to 100%.
- Added custom cooldowns for chill mode users (abilities only, not ultimate).

Notes: 
Relax v1.6.0 changed the entire codebase from using skip-if to if-else statements. If-else statements are one of the basic programming block which is used to control of the flow. Unfortunately, this wasn't available when workshop was introduced back last year. The alternate to that was skip-if statements which worked like go to statements if that makes sense. Because of this, a chunk of code was written using skip ifs. On updating those portions to use if-else statements, it broke the lobby as soon as someone, having an action with if-else running, was swapped to a different team. Since the entire 1.6.x series of updates was discarded, the feature of Emote Sync is gone as of now.